// src/http/models/modelsAvatarsListResponse.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * API RMI
 * API para gerenciamento de dados de cidadãos do Rio de Janeiro, incluindo autodeclaração de informações e verificação de contato.
 * OpenAPI spec version: 1.0
 */
import type { ModelsAvatarResponse } from './modelsAvatarResponse'

export interface ModelsAvatarsListResponse {
  data?: ModelsAvatarResponse[]
  page?: number
  per_page?: number
  total?: number
  total_pages?: number
}